ui and business logic seperated
