+++ src/Dash/pages/configuration.py
@@ -1,10 +1,7 @@
 import json
-import os
-import sqlite3
 
 import dash_credit_cards as dcs
 import dash_mantine_components as dmc
-import openpay
 from dash import (
     ALL,
     dcc,
@@ -18,19 +15,14 @@ from dash import (
 )
 from dash.exceptions import PreventUpdate
 from dash_iconify import DashIconify
-from dotenv import load_dotenv
 from flask import current_app, session, flash
 from flask_login import logout_user
 
-load_dotenv()
-openpay.api_key = os.getenv("OPENPAY_APIKEY")
-openpay.verify_ssl_certs = os.getenv("OPENPAY_VERIFY_SSL_CERTS") == 'True'
-openpay.merchant_id = os.getenv("OPENPAY_MERCHANT_ID")
-openpay.country = os.getenv("OPENPAY_COUNTRY")
+from src.Dash.services.database import OpenPay
+
+api = OpenPay()
 
 register_page(__name__, name="Configuration", path=current_app.config["URL_CONFIGURATION"])
-
-subscription_data = []
 
 
 def layout(**kwargs):
@@ -42,7 +34,7 @@ def layout(**kwargs):
         return full_layout(None, None, None, {"message": "No OpenPay ID found for the user.",
                                               "error_message": "No OpenPay ID found for the user."})
     else:
-        customer, subscription, plan, error_message = fetch_subscription_and_plan(user.get("openpay_id", ""))
+        customer, subscription, plan, error_message = api.fetch_subscription_and_plan(user.get("openpay_id", ""))
     return full_layout(customer["id"], subscription, plan, error_message)
 
 
@@ -61,7 +53,8 @@ def full_layout(customer_id, subscription, plan, error_message):
             style={"padding": "20px"},
         )
     else:
-        subscription_card_ids = [sub["card"]["id"] for sub in subscription_data["data"] if sub["status"] == "active"]
+        subscription_card_ids = [sub["card"]["id"] for sub in api.subscription_data["data"] if
+                                 sub["status"] == "active"]
         return html.Div(
             [
                 dmc.Title("Customer Configuration", order=2),
@@ -143,7 +136,7 @@ def full_layout(customer_id, subscription, plan, error_message):
                                           "borderRadius": "15px",
                                       }
                                   )
-                                  for card in cards_data
+                                  for card in api.cards_data
                               ]
                               or [html.Div("No cards available")]
                               ),
@@ -179,36 +172,6 @@ def full_layout(customer_id, subscription, plan, error_message):
         )
 
 
-def fetch_subscription_and_plan(openpay_id):
-    try:
-        customer = openpay.Customer.retrieve(openpay_id)
-        global subscription_data
-        subscription_data = customer.subscriptions.all()
-
-        global cards_data
-        cards_data = fetch_credit_cards(customer)
-
-        active_subscription = None
-        for subscription in subscription_data["data"]:
-            if subscription.get("status") == "active":
-                active_subscription = subscription
-                break
-
-        if active_subscription:
-            plan = openpay.Plan.retrieve(active_subscription["plan_id"])
-            return customer, active_subscription, plan, {"message": None, "error_message": None}
-        else:
-            return customer, None, None, {"message": "No active subscriptions found.", "error_message": None}
-    except openpay.error.OpenpayError as e:
-        print("Error fetching subscription details:", e)
-        return None, None, None, {"message": None, "error_message": "Error fetching subscription details: {}".format(e)}
-
-
-def fetch_credit_cards(customer):
-    cards = customer.cards.all()
-    return cards["data"]
-
-
 @callback(
     Output("credit_cards_display", "children", allow_duplicate=True),
     Output("credit_card_deleted_message", "children"),
@@ -223,19 +186,17 @@ def delete_credit_card_callback(n_clicks_list):
     triggered_id = json.loads(triggered_id_str)
     card_id = triggered_id["card_id"]
     if card_id:
-        global cards_data
-        cards_data = [card for card in cards_data if card.get("id") != card_id]
-        if not delete_card_from_openpay(triggered_id["customer_id"], card_id):
+        api.cards_data = [card for card in api.cards_data if card.get("id") != card_id]
+        if not api.delete_card_from_openpay(triggered_id["customer_id"], card_id):
             return no_update, "Error: deleting card from OpenPay"
 
         # Handle the case where there are no subscriptions
-        global subscription_data
-        if subscription_data and "data" in subscription_data:
-            subscription_card_ids = [sub["card"]["id"] for sub in subscription_data["data"] if
+        if api.subscription_data and "data" in api.subscription_data:
+            subscription_card_ids = [sub["card"]["id"] for sub in api.subscription_data["data"] if
                                      sub["status"] == "active"]
         else:
             subscription_card_ids = []
-        if cards_data:
+        if api.cards_data:
             updated_card_displays = [
                 dmc.Paper(
                     [
@@ -274,46 +235,11 @@ def delete_credit_card_callback(n_clicks_list):
                         "borderRadius": "15px",
                     },
                 )
-                for card in cards_data
+                for card in api.cards_data
             ]
             return updated_card_displays, "Card deleted successfully!"
         else:
             return [html.Div("No cards available")], ""
-
-
-def delete_card_and_purchases_from_db(card_id):
-    try:
-        conn = sqlite3.connect(current_app.config["DATABASE_PATH"])
-        cursor = conn.cursor()
-
-        delete_purchases_query = """DELETE FROM purchases WHERE card_id = ?"""
-        cursor.execute(delete_purchases_query, (card_id,))
-        delete_card_query = """DELETE FROM cards WHERE card_id = ?"""
-        cursor.execute(delete_card_query, (card_id,))
-
-        conn.commit()
-        conn.close()
-        return True
-    except Exception as e:
-        print("Error deleting card and purchases from database:", e)
-        return False
-
-
-def delete_card_from_openpay(customer_id, card_id):
-    try:
-        customer = openpay.Customer.retrieve(customer_id)
-        card = customer.cards.retrieve(card_id)
-        card.delete()
-        # Delete card and purchases from the database
-        if not delete_card_and_purchases_from_db(card_id):
-            raise Exception("Failed to delete card and purchases from the database.")
-        return True
-    except openpay.error.OpenpayError as e:
-        print("Error deleting card from OpenPay:", e)
-        return False
-    except Exception as e:
-        print(e)
-        return False
 
 
 @callback(
@@ -330,8 +256,8 @@ def delete_user_callback(n_clicks):
         if openpay_id is None or openpay_id == "":
             return "No OpenPay ID found for the user.", no_update
         # Delete OpenPay account first
-        if delete_account_from_openpay(openpay_id):
-            if delete_user(email, openpay_id, True, True, True):
+        if api.delete_account_from_openpay(openpay_id):
+            if api.delete_user(email, openpay_id, True, True, True):
                 logout_user()
                 flash("Account deleted successfully.", "warning")
                 return "Account deleted successfully.", current_app.config["URL_SIGNUP"]
@@ -340,40 +266,6 @@ def delete_user_callback(n_clicks):
         else:
             return "Failed to delete OpenPay account. Please try again later.", no_update
     raise PreventUpdate
-
-
-def delete_user(email=None, openpay_id=None, delete_user=False, delete_cards=False, delete_purchases=False):
-    if not any([delete_user, delete_cards, delete_purchases]):
-        print("No deletion action specified.")
-        return False
-    try:
-        conn = sqlite3.connect(current_app.config["DATABASE_PATH"])
-        cursor = conn.cursor()
-
-        if delete_user and email:
-            # Remove user from the users table
-            delete_user_query = """DELETE FROM users WHERE email = ?"""
-            cursor.execute(delete_user_query, (email,))
-            print(f"User {email} deleted.")
-
-        if delete_cards and openpay_id:
-            # Remove cards for that specific customer
-            delete_cards_query = """DELETE FROM cards WHERE customer_id = ?"""
-            cursor.execute(delete_cards_query, (openpay_id,))
-            print(f"Cards for customer {openpay_id} deleted.")
-
-        if delete_purchases and openpay_id:
-            # Remove purchases for that specific customer
-            delete_purchases_query = """DELETE FROM purchases WHERE customer_id = ?"""
-            cursor.execute(delete_purchases_query, (openpay_id,))
-            print(f"Purchases for customer {openpay_id} deleted.")
-
-        conn.commit()
-        conn.close()
-        return True
-    except Exception as e:
-        print("Error deleting user details:", e)
-        return False
 
 
 @callback(
@@ -389,8 +281,8 @@ def delete_openpay_account_callback(n_clicks):
         email = user_session.get("email", "")
         openpay_id = user_session.get("openpay_id", "")
         if openpay_id:
-            if delete_account_from_openpay(openpay_id):
-                delete_user(email, openpay_id, False, True, True)
+            if api.delete_account_from_openpay(openpay_id):
+                api.delete_user(email, openpay_id, False, True, True)
                 return "Openpay account and associated data deleted successfully.", dcc.Location(
                     pathname=current_app.config["URL_CONFIGURATION"],
                     id="redirect"), {"refresh": True}
@@ -399,25 +291,6 @@ def delete_openpay_account_callback(n_clicks):
         else:
             return "No OpenPay ID found for the user.", no_update, no_update
     raise PreventUpdate
-
-
-def delete_account_from_openpay(customer_id):
-    try:
-        customer = openpay.Customer.retrieve(customer_id)
-        subscriptions = customer.subscriptions.all()
-        for subscription in subscriptions["data"]:
-            if subscription["status"] == "active":
-                subscription.delete()
-        cards = customer.cards.all()
-        for card in cards["data"]:
-            card.delete()
-
-        # Delete the customer account itself
-        # openpay.Customer.delete(customer_id)
-        return True
-    except openpay.error.OpenpayError as e:
-        print("Error deleting account from OpenPay:", e)
-        return False
 
 
 @callback(
@@ -431,7 +304,7 @@ def delete_account_from_openpay(customer_id):
 def refresh_data(store_data):
     if store_data and store_data.get("refresh"):
         openpay_id = session["user"].get("openpay_id", "")
-        customer, subscription, plan, message = fetch_subscription_and_plan(openpay_id)
+        customer, subscription, plan, message = api.fetch_subscription_and_plan(openpay_id)
 
         if plan:
             plan_name = f"Plan Name: {plan['name']}"
@@ -440,8 +313,8 @@ def refresh_data(store_data):
             plan_name = "No subscription available"
             plan_price = ""
 
-        if cards_data:
-            subscription_card_ids = [sub["card"]["id"] for sub in subscription_data["data"] if
+        if api.cards_data:
+            subscription_card_ids = [sub["card"]["id"] for sub in api.subscription_data["data"] if
                                      sub["status"] == "active"]
             card_displays = [
                 dmc.Paper(
@@ -481,7 +354,7 @@ def refresh_data(store_data):
                         "borderRadius": "15px",
                     },
                 )
-                for card in cards_data
+                for card in api.cards_data
             ]
             return card_displays, plan_name.plan_price, "Openpay account and associated data deleted successfully."
         else:
